Use legacy number series for MHR number generation.
